fix(economy): use int instead of long for TradeCompletedEvent.totalPrice
Matches the spec and keeps consistency with the pricing formula output type.

+++ core/src/main/java/com/galacticodyssey/economy/events/TradeCompletedEvent.java
@@ -5,11 +5,11 @@ public final class TradeCompletedEvent {
     public final String commodityId;
     public final int quantity;
     public final int unitPrice;
-    public final long totalPrice;
+    public final int totalPrice;
     public final boolean isBuy;
 
     public TradeCompletedEvent(String stationId, String commodityId, int quantity,
-                                int unitPrice, long totalPrice, boolean isBuy) {
+                                int unitPrice, int totalPrice, boolean isBuy) {
         this.stationId = stationId;
         this.commodityId = commodityId;
         this.quantity = quantity;
